refactor(pr2): drop residual D content from agent CLI
Trim missed agent-side D content — the JWT enrollment refactor lives in
its own follow-up PR, so PR2 should not carry any of it:

- enrollment.rs: restore EnrollResponse::quic_endpoint and the original
  enroll_agent / persist_enrollment_response signatures (no extra
  quic_endpoint or advertise_domains parameters). Drop the
  EnrollmentJwtClaims::jet_quic_endpoint claim — the enrollment JWT
  carries gw_url / agent_name only.
- main.rs: drop --quic-endpoint CLI flag, drop UpCommand::quic_endpoint,
  drop the JWT jet_quic_endpoint extraction, restore the two-arg-shorter
  service-mode signature, restore the inline cli tests module.
- cli_tests.rs: removed (the tests are back inline in main.rs at the
  master state).
